easy/requirements.txt
tensorflow==1.13.1
regex==2020.11.13
networkx==2.5
Keras==2.2.4
scipy==1.5.2
tqdm==4.55.1
torch_geometric==1.6.3
transformers==4.1.1
torch==1.7.1
torch_scatter==2.0.5
git+https://github.com/fwilliams/fml
torch_sparse==0.6.8
pickle5==0.0.11
numpy==1.19.2
graph_tool==2.11
python_igraph==0.8.3
python_Levenshtein==0.12.2
scikit_learn==0.24.1
